libc: use usize for sbrk() increment

+++ sys/include/libc.h
@@ -714,7 +714,7 @@ extern	long	read(int, void*, long);
 extern	long	readn(int, void*, long);
 extern	long	readv(int, IOchunk*, int);
 extern	int	remove(char*);
-extern	void*	sbrk(ulong);
+extern	void*	sbrk(usize);
 extern	long	oseek(int, long, int);
 extern	vlong	seek(int, vlong, int);
 extern	void*	segattach(int, char*, void*, ulong);
